Move routes into separate testing files, otherwise the test_routes file would be too long
