+++ testing/tests/api/test.py
@@ -1,3 +1,4 @@
+import pytest
 from flask import Flask, request, jsonify
 import json
 import requests
