+++ testing/tests/api/test_getindustry.py
@@ -1,0 +1,10 @@
+import pytest
+from unittest.mock import patch
+from flask import Flask, jsonify
+from index import app 
+
+##########################################################################
+# /GETINDUSTRY ROUTE
+# THE FOLLOWING TESTS CHECK THE /GETINDUSTRY ROUTE
+##########################################################################
+
